% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pam50_correlation_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{Generate the PAM50 correlation coefficient plots}
\usage{
cor_plot(p50, id, name)
}
\arguments{
\item{p50}{The PAM50 results text file}

\item{id}{The ID of the samples to extract}

\item{name}{The name to represent the samples}
}
\description{
Generate the PAM50 correlation coefficient plots
}

