% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subtype_functions.R
\name{heatmap_col_colours}
\alias{heatmap_col_colours}
\title{Define the column colours for the heatmap.}
\usage{
heatmap_col_colours(m)
}
\arguments{
\item{m}{A data frame with gene symbols as row names.}
}
\description{
Define the column colours for the heatmap.
}

