% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subtype_functions.R
\name{makeHeatmap}
\alias{makeHeatmap}
\title{Generate the heatmaps.}
\usage{
makeHeatmap(m, title, pam.res)
}
\arguments{
\item{m}{A data frame with gene symbols as row names.}

\item{title}{The title for the heatmap}

\item{pam.res}{The PAM50 results file}
}
\description{
Generate the heatmaps.
}

