% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subtype_functions.R
\name{run_basal}
\alias{run_basal}
\title{Run the subtype analysis}
\usage{
run_basal(d, f, s)
}
\arguments{
\item{inputDir}{The directory containing the input file}

\item{inputFile}{A dataframe of expression data with row names as gene symbols and column names as unique sample IDs}

\item{short}{A short name for the analysis}
}
\description{
Run the subtype analysis
}

