% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subtype_functions.R
\name{run_basalt}
\alias{run_basalt}
\title{Run the subtype analysis}
\usage{
run_basalt(outDir, inputFile, short, pamCC = 0.75, samp_num = 5,
  remove_lc = TRUE)
}
\arguments{
\item{outDir}{The directory containing for the output}

\item{inputFile}{The full path to a dataframe of expression data with row names as gene symbols and column names as unique sample IDs}

\item{short}{A short name for the analysis}

\item{pamCC}{(default=0.75) Correlation confidence cutoff used to assign a PAM50 subtype as Low Confidence (LC)}

\item{samp_num}{(default=5) Minumum percentage of samples containing value > 0 per gene}

\item{remove_lc}{(default TRUE) Remove the samples with a corerrelation confidence less than pamCC}
}
\description{
Run the subtype analysis
}

