% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subtype_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Define the row colours for the heatmap.}
\usage{
plot_summary(m, outDir)
}
\arguments{
\item{m}{The final subtype summary dataframe}
}
\description{
Define the row colours for the heatmap.
}

